// migrations/snapshots/fbd2995f4d9662d65e2c4f1a5657aff1786aeb0a3ed76cd28e83e19039e0bb7a/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Bookings: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly fieldId: CodecTypes['pg/uuid@1']['output'];
      readonly bookingDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly totalPrice: CodecTypes['pg/int4@1']['output'];
      readonly status: 'PENDING' | 'CONFIRMED' | 'CANCELLED' | 'COMPLETED' | 'EXPIRED';
      readonly paymentStatus: 'UNPAID' | 'PAID' | 'FAILED' | 'REFUNDED' | 'EXPIRED';
      readonly notes: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly durationHours: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly BookingTimeSlots: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly bookingId: CodecTypes['pg/uuid@1']['output'];
      readonly startTime: CodecTypes['pg/time-temporal@1']['output'];
      readonly endTime: CodecTypes['pg/time-temporal@1']['output'];
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly fieldId: CodecTypes['pg/uuid@1']['output'];
      readonly bookingDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly FieldImages: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly fieldId: CodecTypes['pg/uuid@1']['output'];
      readonly imageUrl: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly FieldOperatingHours: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly fieldId: CodecTypes['pg/uuid@1']['output'];
      readonly dayOfWeek: CodecTypes['pg/int4@1']['output'];
      readonly openTime: CodecTypes['pg/time-temporal@1']['output'];
      readonly closeTime: CodecTypes['pg/time-temporal@1']['output'];
    };
    readonly Fields: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<100>;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly pricePerHour: CodecTypes['pg/int4@1']['output'];
      readonly imageUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly slug: Varchar<120>;
      readonly location: CodecTypes['pg/text@1']['output'] | null;
      readonly _type: Varchar<50> | null;
      readonly capacity: CodecTypes['pg/int4@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Transactions: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly bookingId: CodecTypes['pg/uuid@1']['output'];
      readonly orderId: Varchar<255>;
      readonly grossAmount: CodecTypes['pg/int4@1']['output'];
      readonly paymentMethod:
        'QRIS' | 'BANK_TRANSFER' | 'E_WALLET' | 'CREDIT_CARD' | 'VA' | 'CASH' | null;
      readonly transactionStatus: 'PENDING' | 'SETTLEMENT' | 'EXPIRE' | 'CANCEL' | 'DENY';
      readonly snapToken: CodecTypes['pg/text@1']['output'] | null;
      readonly paidAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly provider: 'MIDTRANS' | 'XENDIT' | 'MANUAL';
      readonly paymentUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly expiredAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Users: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<100>;
      readonly email: Varchar<255>;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: 'ADMIN' | 'USER';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly phone: Varchar<20> | null;
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Bookings: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly fieldId: CodecTypes['pg/uuid@1']['input'];
      readonly bookingDate: CodecTypes['pg/date-temporal@1']['input'];
      readonly totalPrice: CodecTypes['pg/int4@1']['input'];
      readonly status: 'PENDING' | 'CONFIRMED' | 'CANCELLED' | 'COMPLETED' | 'EXPIRED';
      readonly paymentStatus: 'UNPAID' | 'PAID' | 'FAILED' | 'REFUNDED' | 'EXPIRED';
      readonly notes: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly durationHours: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly BookingTimeSlots: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly bookingId: CodecTypes['pg/uuid@1']['input'];
      readonly startTime: CodecTypes['pg/time-temporal@1']['input'];
      readonly endTime: CodecTypes['pg/time-temporal@1']['input'];
      readonly price: CodecTypes['pg/int4@1']['input'];
      readonly fieldId: CodecTypes['pg/uuid@1']['input'];
      readonly bookingDate: CodecTypes['pg/date-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly FieldImages: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly fieldId: CodecTypes['pg/uuid@1']['input'];
      readonly imageUrl: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly FieldOperatingHours: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly fieldId: CodecTypes['pg/uuid@1']['input'];
      readonly dayOfWeek: CodecTypes['pg/int4@1']['input'];
      readonly openTime: CodecTypes['pg/time-temporal@1']['input'];
      readonly closeTime: CodecTypes['pg/time-temporal@1']['input'];
    };
    readonly Fields: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly pricePerHour: CodecTypes['pg/int4@1']['input'];
      readonly imageUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly slug: CodecTypes['sql/varchar@1']['input'];
      readonly location: CodecTypes['pg/text@1']['input'] | null;
      readonly _type: CodecTypes['sql/varchar@1']['input'] | null;
      readonly capacity: CodecTypes['pg/int4@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Transactions: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly bookingId: CodecTypes['pg/uuid@1']['input'];
      readonly orderId: CodecTypes['sql/varchar@1']['input'];
      readonly grossAmount: CodecTypes['pg/int4@1']['input'];
      readonly paymentMethod:
        'QRIS' | 'BANK_TRANSFER' | 'E_WALLET' | 'CREDIT_CARD' | 'VA' | 'CASH' | null;
      readonly transactionStatus: 'PENDING' | 'SETTLEMENT' | 'EXPIRE' | 'CANCEL' | 'DENY';
      readonly snapToken: CodecTypes['pg/text@1']['input'] | null;
      readonly paidAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly provider: 'MIDTRANS' | 'XENDIT' | 'MANUAL';
      readonly paymentUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly expiredAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Users: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly email: CodecTypes['sql/varchar@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: 'ADMIN' | 'USER';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly phone: CodecTypes['sql/varchar@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;